djangobackend/requirements.txt
amqp==5.2.0
aniso8601==7.0.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
cryptography==42.0.8
DateTime==5.5
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
Django
django-allauth==0.60.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
environ==1.0
filelock==3.15.4
graphene==2.1.9
graphene-django==2.2.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==23.0.0
idna==3.7
kombu==5.4.0
node==1.2.2
NOV==1.0
novu==1.14.0
oauthlib==3.2.2
odict==1.9.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
plumber==1.7
promise==2.3
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
Rx==1.6.3
setuptools==71.0.4
singledispatch==4.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
text-unidecode==1.3
timedelta==2020.12.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.4.post2
zope.lifecycleevent==5.0
zope.proxy==5.2
